pyproject.toml
[tool.poetry]
name = "azure-devops-slack"
version = "0.0.1"
description = "Slack notifications from Azure DevOps pull requests"
authors = ["Denis <denis@fbjorn.cc>"]
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
Flask = "^2.2.2"
pydantic = "^1.10.4"
python-dotenv = "^0.21.1"
functions-framework = "^3.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
syrupy = "^4.5.0"
ruamel-yaml = "^0.17.32"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
